[vm/concurrency] Avoid holding the program lock when installing new TTS
When intalling a new TTS stub two fields have to be updated:

    - AbstractType::type_test_stub
    - AbstractType::type_test_stub_entry_point_

Right now the code takes the program lock when updating those two
fields, which is very heavy weight, since this happens very often and
therefore the lock will be contended.

Instead we'll solve the write-write races by ensuring they are applied
using atomics in (any) order - though ensuring that they will be in
sync.

It is fine for them to be out-of sync for a short time, as long as we're
guaranteed that whenever we hit a gc safepoint they are in sync (because GC
might collect old TTS stubs that are not referenced anymore - i.e. we
need to avoid situations where the cached entrypoint is old but the code
pointer is new)

There can be multiple mutators computing and installing a TTS at the same
time, they will try to install in sequence.

=> This reduces the average waiting time on write access to
   program lock by 3x

Fixes https://github.com/dart-lang/sdk/issues/46124
Issue https://github.com/dart-lang/sdk/issues/46252

TEST=Existing test coverage.

Change-Id: Id1ed29adeb26fa93494e8710206fd680d427bd34
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/205520
Commit-Queue: Martin Kustermann <kustermann@google.com>
Reviewed-by: Slava Egorov <vegorov@google.com>