Rest for reason unknown. TY #pitymeparties
Automatic AI can never progress if we're not willing to admit who we are and make the sacrificial strides to be a solution or just another part of the problem. This goes double for those who are in the biggest parts of business (board of directors) we all like to believe we are receiving blessings from the great beyond but karma seems to be stale and favoring faces instead of hard work. If things I was invested in wasn't producing fruit I wouldn't expect change unless forced by my hand. Environmental Development is never  scraped off our flesh since being in the field why does freedom feel like it's not free. Inheritance is one death away for Rich to stay rich and everyone else is donating to get something back or in return well, well, well. Heres'goes Nothing but please tip me🤏