[cfe] Use correct library when building origin outline expressions
When constructors are synthesized for mixin applications, the default
values and initializers of the original constructor needs to be copied
over for const constructors. Since the original default values and
initializers may not have been created yet, when these are required
for the synthesized constructor, the synthesized constructor triggers
the computation on the original constructor. This triggering wrongly
used the library of the synthesized constructor and not the original
constructor as the context for this computation, thus not handling
privacy correctly.

Closes #46581

Change-Id: If0c5129427c7e7c8d6434bbaa96b4df559dccc85
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/206200
Commit-Queue: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>