[VM] Change internal name for FfiNatives.
Changes the naming format for synthetic variables used for
@FfiNatives.
The name was based on the native function, but since multiple
FfiNatives can reference the same native function, this can
result in name clashes.
Instead the name is now based on the name of the associated
function, which cannot have a name clashing with another function
by the same name.

TEST=CQ
Change-Id: Id829ff3f207ab11e83f3194b8410f4be297623c4
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/206374
Commit-Queue: Clement Skau <cskau@google.com>
Reviewed-by: Tess Strickland <sstrickl@google.com>