[vm/concurrency] Remove assertion about deoptimization counts
When doing optimizing compilations on the mutator thread (e.g. due to
--deterministic or --no-background-compilation) we can have situations
in which one thread is starting or in the middle of optimizing a
function while another has just deoptimized it, reaching the
deoptimization counter threshold.

We just remove this assertion which doesn't hold anymore with
--enable-isolate-groups enabled.

Fixes https://github.com/dart-lang/sdk/issues/46622

TEST=Fixes flaky assertion hit in vm/dart_2/isolates/regress_46539_test

Change-Id: I9812e3c6500113b7d948f96dc4cfcee607f56d39
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/207023
Reviewed-by: Tess Strickland <sstrickl@google.com>
Commit-Queue: Martin Kustermann <kustermann@google.com>