[vm/concurrency] Remove unix_socket_test.dart from iso-stress builder
The test causes flaky issues when run in parallel, it may also have
client/server races in them.

The flaky failures happen due to a socket write resulting in a "broken
pipe" exception. Normally that means the other end of the socket was
closed whent the write happens.

An attempt in [0] at fixing one of the issues seem to be unsuccessful,
since the error still happens, namely in this code:

  final serverContinue = Completer();
  final clientContinue = Completer();
  server.listen((s) async {
    await serverContinue.future;
    clientContinue.complete();
  });

  final socket = await Socket.connect(address, server.port);
  socket.write(" socket content");  // <-- This write causes a broken pipe exception!
  serverContinue.complete();
  await clientContinue.future;

It's unclear why that happens, since the server side certainly doesn't
close it's end of the socket.

Since this is causing "iso-stress" builder to be failing, I'll remove
this test from the stress test.

Issue https://github.com/dart-lang/sdk/issues/36097

TEST=Removing test from "iso-stress" builder

Change-Id: I1f9fc130b7408d586b55699487baa3c293403217
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/207134
Reviewed-by: Martin Kustermann <kustermann@google.com>