[analyzer_plugin] Support latest analyzer in pubspec
As the analyzer_plugin package is always built with the latest analyzer,
its pubspec should also report compatibility with the latest released
version.

This changes the analyzer constraint to `^2.0.0` which matches the
current version on pub.
It would be great if the next analyzer_plugin release could support the
latest analyzer.

Change-Id: I2a7b129f199bd80e6ca552821feb43b29193d2c0
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/207302
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>