[cfe] Trigger constructor parameter inference on constructor tear off
The type of a constructor tear off is dependent on the type of
the constructor. Since the constructor type might be inferred,
for instance for initializing formals, we need to trigger this inference
before computing the static type of a constructor tear off.

This needs to be further expanded to lowerings of tear offs. This will
be done in a follow-up.

Change-Id: I321c013ec60e2f17b99a41c8f9debc2f564462f3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/207940
Reviewed-by: Dmitry Stefantsov <dmitryas@google.com>