Move shouldIgnoreUndefined logic into LibraryElementImpl.
This logic was previously in LibraryScope and ScopeExtension.  It
really isn't related to scopes, though, since all it does is to look
through the library's imports and parts.  (The only use it made of
scopes was to locate the library element).

Separating this code from scopes will facilitate some other changes
I'm working on to streamline the use of scopes in the analyzer.

Change-Id: Ic45e3382628a433a96cc7ecf751b4df5d681b42a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/208522
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>