Improve `String.substring` documentation
This PR provides additional info on the `substring` method, as it will throw a `RangeError` in the following cases:

* `start` is negative;
* `end` is negative;
* `start` is greater than `length`;
* `end` is greater than `length`;
* `start` is greater than `end`.

Although some of them are a little obvious, it's better to be explicit about the expectations. One could expect that, for example, that `'abc'.substring(0, 4)` would return `'abc'` instead of throwing.

Closes https://github.com/dart-lang/sdk/pull/46640
https://github.com/dart-lang/sdk/pull/46640

GitOrigin-RevId: d878a4cfbd1eedc7b7e7fc5d6860321c87317346
Change-Id: If3c55b294e2fefd83b7ee5cdeed6478dbe2c7448
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/207200
Reviewed-by: Lasse R.H. Nielsen <lrn@google.com>
Commit-Queue: Lasse R.H. Nielsen <lrn@google.com>