[cfe] Cleanup detection of cyclic redirecting factories
This is a follow-up to [1]. Since that change we started detecting
cyclic factories earlier, so cycle detection can be removed from the
later stage of resolving calls to redirecting factories.

Similarly, detection of redirecting factories targeting generative
constructors of abstract classes was also moved to an earlier stage
and can be cleaned up.

[1] https://dart-review.googlesource.com/c/sdk/+/208681

Change-Id: Ib3d061438c7a898424ffd73f18e2a289a1f987fd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/209000
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>