Move some logic from ScopedVisitor to ResolverVisitor.
Of the two classes that inherit from ScopedVisitor (ResolverVisitor
and ScopeResolverVisitor), only ResolverVisitor needs to know about
the currently enclosing class, extension, and function.  So we move
this tracking logic to the ResolverVisitor itself.  This will help
pave the way for a cleaner separation between the ResolverVisitor and
ScopeResolverVisitor.

(The ScopeResolverVisitor *does* need to know the currently enclosing
closure, so that it can detect improper break/continue statements and
so that it can detect when a local variable is mutated in a closure,
but the tracking logic needed for this is much simpler than what the
ResolverVisitor needs for the "enclosing function".  So a new
_enclosingClosure member has been added to ScopeResolverVisitor to
track this.)

Change-Id: I768c15ff4ffcabf564388f5471a6e8f88e677148
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/209665
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>