Refactor scope lookup logic in preparation for consolidating scope analysis to a single visitor.
This CL refactors the scope lookup logic for simple identifiers so
that it happens in three phases:

- The actual lookup, handled by Scope.lookup

- Choosing which element is being referred to, handled by the
  LexicalLookup class

- If lexical lookup failed, falling back on implicit `this` lookup,
  handled by the ThisLookup class

In a future CL, this will allow us to consolidate the first phase,
along with all other scope-related operations, into a single visitor
so that the ResolverVisitor no longer needs to maintain scopes.  This
should simplify the analyzer's scoping logic and help prepare for
sharing more code with the CFE (since it will make the analyzer's
visitors more closely parallel the analysis phases of the CFE).

Change-Id: I0c8a2aa60bfc5a935b6060adf64e5a4504bc268f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/209664
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>