Create a new base class for resolution visitors.
The new class, ResolverBase, will eventually take the place of
ScopedVisitor as the common base class between ResolverVisitor and
ScopeResolverVisitor.  Unlike ScopedVisitor, it won't maintain any
scope information (since we are migrating all scope handling out of
ResolverVisitor); all it needs to do is keep track of some information
that is needed by all resolution stages, such as the library element,
the type provider, and the error reporter.

Change-Id: I00f80457d368489d6df9736713fd3caddd5f917f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/209720
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>